+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Added
 
 - [TD-4540] Add approval workflow for quality implementations
+### Fixed
+
+- [TD-4852] Ruleless implementation weren't being linked with data structures on
+  creation
 
 ## [4.44.0] 2022-05-23
 
@@ -70,7 +74,7 @@
 
 - [TD-4721] User with permissions can list sources
 
-# Add
+### Added
 
 - [TD-4695] Structure bulk update was not reindexing structures with updated
   notes
@@ -80,7 +84,7 @@
 
 ## [4.41.1] 2022-04-04
 
-# Added
+### Added
 
 - [TD-4126] Support querying the jobs of a source
 - [TD-4450] GraphQL mutations for sources
@@ -110,7 +114,7 @@
 - [TD-4580] Added migration to fix implementations without alias
 - [TD-4623] Added multiple domains support for grant request approvals
 
-## Changed
+### Changed
 
 - [TD-4527] Exclude logically deleted sources from GraphQL response